ai-readme-cli/requirements.txt
click>=8.0.0
inquirer>=2.10.0
rich>=13.0.0
google-generativeai>=0.8.0
gitpython>=3.1.0
pyyaml>=6.0